src/system/requirements.txt
asn1crypto==0.24.0
beautifulsoup4==4.7.1
browsermob-proxy==0.8.0
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
Click==7.0
cryptography==2.4.2
dnspython==1.16.0
future==0.17.1
geoip2==2.9.0
idna==2.8
lxml==4.2.5
maxminddb==1.4.1
numpy==1.15.4
oauthlib==2.1.0
pandas==0.23.4
pkg-resources==0.0.0
praw==6.0.0
prawcore==1.0.0
pyasn==1.6.0b1
pycparser==2.19
pyOpenSSL==18.0.0
PySocks==1.6.8
python-dateutil==2.7.5
python-whois==0.7.1
pytz==2018.9
requests==2.21.0
requests-file==1.4.3
requests-oauthlib==1.0.0
selenium==3.141.0
six==1.12.0
soupsieve==1.7
tldextract==2.2.0
tweepy==3.7.0
unshortenit==0.4.0
update-checker==0.16
urllib3==1.24.1
domaintools_api
scikit-learn
matplotlib
